+++ app/src/main/res/layout/activity_message.xml
@@ -47,6 +47,7 @@
         android:layout_marginEnd="32dp"
         android:layout_marginRight="32dp"
         android:layout_marginBottom="499dp"
+
         android:text="@string/tilbakeTxt"
         android:textColor="@android:color/white"
         android:textSize="18sp"
